requirements.txt
hidapi==0.10.1
psutil==5.8.0
PyQt5-sip==12.8.1
PySide2==5.15.2
shiboken2==5.15.2
